Updated HomeNewsFeed.tsx to really work with Instagram Feed
I found something that allowed me to experiment with React for a bit and so I was able to figure out the syntax needed to make this work. I can't be completely sure if it will work in production but this should at least be a step in the right direction.