Update spotlight changes to remove potential edgecases
The compiler keeps getting mad about edge cases, (even caring about type never which is meant to be an impossible value apparently) so I'm just gonna remove the rest of the potential for edge cases all at once, and hopefully, it'll work. If it doesn't then I'll work at it more.